[cherry-pick] Fix distill codes and docs (#7742)
* fix distill codes and docs

* fix slim distill docs

* fix docs

* add details of docs, test=document_fix
