+++ ppdet/modeling/heads/__init__.py
@@ -35,7 +35,6 @@ from . import retina_head
 from . import ppyoloe_head
 from . import fcosr_head
 from . import ppyoloe_r_head
-from . import ld_gfl_head
 from . import yolof_head
 from . import ppyoloe_contrast_head
 from . import centertrack_head
@@ -63,7 +62,6 @@ from .tood_head import *
 from .retina_head import *
 from .ppyoloe_head import *
 from .fcosr_head import *
-from .ld_gfl_head import *
 from .ppyoloe_r_head import *
 from .yolof_head import *
 from .ppyoloe_contrast_head import *
